+++ src/components/button/buttonBase.tsx
@@ -34,20 +34,20 @@ export function getVisualStates(props: ButtonBaseProps, componentName: string) {
 
 export function getFocusColor(props: ButtonBaseProps) {
     const { mode, color } = props;
-    let focusColor: string = White.primary;
+    let focusColor: string = White.focus;
 
     switch (color) {
         case Color.default:
-            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Blue.primary : White.primary;
+            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Blue.focus : White.focus;
             break;
         case Color.danger:
-            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Red.focus : White.primary;
+            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Red.focus : White.focus;
             break;
         case Color.success:
-            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Green.focus : White.primary;
+            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Green.focus : White.focus;
             break;
         case Color.warning:
-            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Orange.focus : White.primary;
+            focusColor = (mode === Mode.text || mode === Mode.outlined) ? Orange.focus : White.focus;
             break;
     }
 
@@ -60,16 +60,16 @@ export function getWaveColor(props: ButtonBaseProps) {
 
     switch (color) {
         case Color.default:
-            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Blue.wave : White.primary;
+            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Blue.deepLight : White.focus;
             break;
         case Color.danger:
-            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Red.wave : White.primary;
+            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Red.deepLight : White.focus;
             break;
         case Color.success:
-            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Green.wave : White.primary;
+            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Green.deepLight : White.focus;
             break;
         case Color.warning:
-            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Orange.wave : White.primary;
+            waveColor = (mode === Mode.text || mode === Mode.outlined) ? Orange.deepLight : White.focus;
             break;
     }